+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## 1.13.0-b8
+## 1.13.0-b9
 ### 05-Oct-2023
 - 🔥 **Breaking Change** for users of Setup Your Mac prior to `1.13.0` 🔥 
   - Removed `setupYourMacPolicyArrayIconPrefixUrl` (in favor using the fully qualified domain name of the server which hosts your icons)
@@ -10,6 +10,9 @@
 - Added `supportTeamWebsite` (Addresses [Issue No. 97](https://github.com/dan-snelson/Setup-Your-Mac/issues/97); thanks, @theahadub!)
 - Modified the display of support-related information
 - Adjustments to the `wait` flavor of `completionActionOption` (thanks for the heads-up, @Tom!)
+- Added "COMPLETION ACTION:" to all `completionActionOption`-related log entries
+- More adjustments to the display of support-related information
+- Adjustments to `messageOnly` `completionActionOption`
 
 ## 1.12.12
 ### 28-Sep-2023